Refactor our internal representation of a call graph to represent 'borrows from' relationships using edge metadata rather than standalone nodes
