+++ libs/pavex_cli/tests/ui_tests/app_builder/expectations/diagnostics.dot
@@ -7,23 +7,21 @@ digraph "GET /home" {
     5 [ label = "http::Request<hyper::Body>"]
     6 [ label = "<http::Response::<http_body::combinators::BoxBody::<bytes::Bytes, pavex_runtime::Error>> as pavex_runtime::response::IntoResponse>::into_response(http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
     7 [ label = "core::prelude::rust_2015::v1::Result<std::path::PathBuf, app::ExtractPathError<alloc::string::String>> -> app::ExtractPathError<alloc::string::String>"]
-    8 [ label = "app::ExtractPathError<alloc::string::String> -> &app::ExtractPathError<alloc::string::String>"]
-    9 [ label = "app::handle_extract_path_error(&app::ExtractPathError<alloc::string::String>, app::Logger) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
-    10 [ label = "app::logger() -> app::Logger"]
-    11 [ label = "<http::Response::<http_body::combinators::BoxBody::<bytes::Bytes, pavex_runtime::Error>> as pavex_runtime::response::IntoResponse>::into_response(http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
-    12 [ label = "`match`"]
+    8 [ label = "app::handle_extract_path_error(&app::ExtractPathError<alloc::string::String>, app::Logger) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
+    9 [ label = "app::logger() -> app::Logger"]
+    10 [ label = "<http::Response::<http_body::combinators::BoxBody::<bytes::Bytes, pavex_runtime::Error>> as pavex_runtime::response::IntoResponse>::into_response(http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
+    11 [ label = "`match`"]
     1 -> 0 [ ]
     2 -> 0 [ ]
     3 -> 0 [ ]
-    12 -> 7 [ ]
+    11 -> 7 [ ]
     5 -> 4 [ ]
     0 -> 6 [ ]
-    12 -> 3 [ ]
-    7 -> 8 [ ]
-    8 -> 9 [ ]
-    10 -> 9 [ ]
-    9 -> 11 [ ]
-    4 -> 12 [ ]
+    11 -> 3 [ ]
+    7 -> 8 [ label = "&"]
+    9 -> 8 [ ]
+    8 -> 10 [ ]
+    4 -> 11 [ ]
 }
 digraph app_state {
     0 [ label = "crate::ApplicationState(app::HttpClient) -> crate::ApplicationState"]
