+++ libs/pavex_cli/tests/ui_tests/borrow_checker/diamond_can_be_solved_if_we_can_clone/expectations/diagnostics.dot
@@ -2,22 +2,18 @@ digraph "GET /home" {
     0 [ label = "app::handler(app::C, app::D) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
     1 [ label = "app::d(&app::A, app::B) -> app::D"]
     2 [ label = "app::b() -> app::B"]
-    3 [ label = "app::A -> &app::A"]
-    4 [ label = "app::a() -> app::A"]
-    5 [ label = "app::c(app::A, &app::B) -> app::C"]
-    6 [ label = "app::B -> &app::B"]
-    7 [ label = "<http::Response::<http_body::combinators::BoxBody::<bytes::Bytes, pavex_runtime::Error>> as pavex_runtime::response::IntoResponse>::into_response(http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
-    8 [ label = "<app::B as core::clone::Clone>::clone(&app::B) -> app::B"]
+    3 [ label = "app::a() -> app::A"]
+    4 [ label = "app::c(app::A, &app::B) -> app::C"]
+    5 [ label = "<http::Response::<http_body::combinators::BoxBody::<bytes::Bytes, pavex_runtime::Error>> as pavex_runtime::response::IntoResponse>::into_response(http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
+    6 [ label = "<app::B as core::clone::Clone>::clone(&app::B) -> app::B"]
     1 -> 0 [ ]
-    3 -> 1 [ ]
-    4 -> 3 [ ]
-    5 -> 0 [ ]
-    6 -> 5 [ ]
-    2 -> 6 [ ]
-    4 -> 5 [ ]
-    0 -> 7 [ ]
-    6 -> 8 [ ]
-    8 -> 1 [ ]
+    3 -> 1 [ label = "&"]
+    4 -> 0 [ ]
+    2 -> 4 [ label = "&"]
+    3 -> 4 [ ]
+    0 -> 5 [ ]
+    2 -> 6 [ label = "&"]
+    6 -> 1 [ ]
 }
 digraph app_state {
     0 [ label = "crate::ApplicationState() -> crate::ApplicationState"]
