+++ libs/pavex_cli/tests/ui_tests/constructors_and_handlers_can_fail/expectations/diagnostics.dot
@@ -11,55 +11,47 @@ digraph "GET /home" {
     9 [ label = "core::prelude::rust_2015::v1::Result<app::Logger, app::LoggerError> -> app::LoggerError"]
     10 [ label = "core::prelude::rust_2015::v1::Result<std::path::PathBuf, app::ExtractPathError<alloc::string::String>> -> app::ExtractPathError<alloc::string::String>"]
     11 [ label = "<http::Response::<http_body::combinators::BoxBody::<bytes::Bytes, pavex_runtime::Error>> as pavex_runtime::response::IntoResponse>::into_response(http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
-    12 [ label = "app::HandlerError -> &app::HandlerError"]
-    13 [ label = "app::LoggerError -> &app::LoggerError"]
-    14 [ label = "app::ExtractPathError<alloc::string::String> -> &app::ExtractPathError<alloc::string::String>"]
-    15 [ label = "app::handle_extract_path_error(&app::ExtractPathError<alloc::string::String>, app::Logger) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
-    16 [ label = "core::prelude::rust_2015::v1::Result<app::Logger, app::LoggerError> -> app::Logger"]
-    17 [ label = "app::logger() -> core::prelude::rust_2015::v1::Result<app::Logger, app::LoggerError>"]
-    18 [ label = "app::handle_logger_error(&app::LoggerError) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
-    19 [ label = "app::handle_handler_error(&app::HandlerError) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
+    12 [ label = "app::handle_extract_path_error(&app::ExtractPathError<alloc::string::String>, app::Logger) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
+    13 [ label = "core::prelude::rust_2015::v1::Result<app::Logger, app::LoggerError> -> app::Logger"]
+    14 [ label = "app::logger() -> core::prelude::rust_2015::v1::Result<app::Logger, app::LoggerError>"]
+    15 [ label = "app::handle_logger_error(&app::LoggerError) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
+    16 [ label = "app::handle_handler_error(&app::HandlerError) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
+    17 [ label = "<http::Response::<http_body::combinators::BoxBody::<bytes::Bytes, pavex_runtime::Error>> as pavex_runtime::response::IntoResponse>::into_response(http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
+    18 [ label = "core::prelude::rust_2015::v1::Result<app::Logger, app::LoggerError> -> app::LoggerError"]
+    19 [ label = "<http::Response::<http_body::combinators::BoxBody::<bytes::Bytes, pavex_runtime::Error>> as pavex_runtime::response::IntoResponse>::into_response(http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
     20 [ label = "<http::Response::<http_body::combinators::BoxBody::<bytes::Bytes, pavex_runtime::Error>> as pavex_runtime::response::IntoResponse>::into_response(http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
-    21 [ label = "core::prelude::rust_2015::v1::Result<app::Logger, app::LoggerError> -> app::LoggerError"]
+    21 [ label = "app::handle_logger_error(&app::LoggerError) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
     22 [ label = "<http::Response::<http_body::combinators::BoxBody::<bytes::Bytes, pavex_runtime::Error>> as pavex_runtime::response::IntoResponse>::into_response(http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
-    23 [ label = "<http::Response::<http_body::combinators::BoxBody::<bytes::Bytes, pavex_runtime::Error>> as pavex_runtime::response::IntoResponse>::into_response(http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
-    24 [ label = "app::LoggerError -> &app::LoggerError"]
-    25 [ label = "app::handle_logger_error(&app::LoggerError) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
-    26 [ label = "<http::Response::<http_body::combinators::BoxBody::<bytes::Bytes, pavex_runtime::Error>> as pavex_runtime::response::IntoResponse>::into_response(http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
-    27 [ label = "`match`"]
-    28 [ label = "`match`"]
-    29 [ label = "`match`"]
-    30 [ label = "`match`"]
+    23 [ label = "`match`"]
+    24 [ label = "`match`"]
+    25 [ label = "`match`"]
+    26 [ label = "`match`"]
     1 -> 0 [ ]
     2 -> 0 [ ]
-    28 -> 9 [ ]
+    24 -> 9 [ ]
     4 -> 0 [ ]
-    29 -> 10 [ ]
+    25 -> 10 [ ]
     6 -> 5 [ ]
-    27 -> 8 [ ]
-    27 -> 7 [ ]
-    28 -> 2 [ ]
-    29 -> 4 [ ]
+    23 -> 8 [ ]
+    23 -> 7 [ ]
+    24 -> 2 [ ]
+    25 -> 4 [ ]
     7 -> 11 [ ]
-    8 -> 12 [ ]
-    9 -> 13 [ ]
-    10 -> 14 [ ]
-    14 -> 15 [ ]
-    16 -> 15 [ ]
-    30 -> 21 [ ]
-    13 -> 18 [ ]
-    12 -> 19 [ ]
-    15 -> 20 [ ]
-    30 -> 16 [ ]
-    18 -> 22 [ ]
-    19 -> 23 [ ]
-    21 -> 24 [ ]
-    24 -> 25 [ ]
-    25 -> 26 [ ]
-    0 -> 27 [ ]
-    3 -> 28 [ ]
-    5 -> 29 [ ]
-    17 -> 30 [ ]
+    10 -> 12 [ label = "&"]
+    13 -> 12 [ ]
+    26 -> 18 [ ]
+    9 -> 15 [ label = "&"]
+    8 -> 16 [ label = "&"]
+    12 -> 17 [ ]
+    26 -> 13 [ ]
+    15 -> 19 [ ]
+    16 -> 20 [ ]
+    18 -> 21 [ label = "&"]
+    21 -> 22 [ ]
+    0 -> 23 [ ]
+    3 -> 24 [ ]
+    5 -> 25 [ ]
+    14 -> 26 [ ]
 }
 digraph app_state {
     0 [ label = "crate::ApplicationState(app::HttpClient) -> crate::ApplicationState"]
