+++ libs/pavex_cli/tests/ui_tests/non_static_methods_are_supported/expectations/diagnostics.dot
@@ -1,15 +1,13 @@
 digraph "GET /home" {
     0 [ label = "app::Streamer::stream_file(&app::Streamer, app::Logger) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
     1 [ label = "app::LoggerFactory::logger(&app::LoggerFactory) -> app::Logger"]
-    2 [ label = "&app::LoggerFactory"]
-    3 [ label = "app::Streamer -> &app::Streamer"]
-    4 [ label = "app::Streamer::new() -> app::Streamer"]
-    5 [ label = "<http::Response::<http_body::combinators::BoxBody::<bytes::Bytes, pavex_runtime::Error>> as pavex_runtime::response::IntoResponse>::into_response(http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
+    3 [ label = "app::Streamer::new() -> app::Streamer"]
+    4 [ label = "<http::Response::<http_body::combinators::BoxBody::<bytes::Bytes, pavex_runtime::Error>> as pavex_runtime::response::IntoResponse>::into_response(http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
+    5 [ label = "&app::LoggerFactory"]
     1 -> 0 [ ]
-    2 -> 1 [ ]
-    3 -> 0 [ ]
-    4 -> 3 [ ]
-    0 -> 5 [ ]
+    3 -> 0 [ label = "&"]
+    0 -> 4 [ ]
+    5 -> 1 [ ]
 }
 digraph app_state {
     0 [ label = "crate::ApplicationState(app::LoggerFactory) -> crate::ApplicationState"]
