+++ libs/pavex_cli/tests/ui_tests/references_to_constructible_types_are_allowed/expectations/app.rs
@@ -100,8 +100,8 @@ pub async fn route_handler_0(
 > {
     let v1 = app::request_scoped();
     let v3 = {
-        let v2 = &app::transient();
-        app::stream_file(v0, &v1, v2)
+        let v2 = app::transient();
+        app::stream_file(v0, &v1, &v2)
     };
     <http::Response<
         http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>,
