+++ libs/pavex_cli/tests/ui_tests/references_to_constructible_types_are_allowed/expectations/diagnostics.dot
@@ -1,17 +1,13 @@
 digraph "GET /home" {
     0 [ label = "app::stream_file(&app::Singleton, &app::RequestScoped, &app::Transient) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
-    1 [ label = "app::Transient -> &app::Transient"]
-    2 [ label = "app::transient() -> app::Transient"]
-    3 [ label = "app::RequestScoped -> &app::RequestScoped"]
-    4 [ label = "app::request_scoped() -> app::RequestScoped"]
+    1 [ label = "app::transient() -> app::Transient"]
+    2 [ label = "app::request_scoped() -> app::RequestScoped"]
+    4 [ label = "<http::Response::<http_body::combinators::BoxBody::<bytes::Bytes, pavex_runtime::Error>> as pavex_runtime::response::IntoResponse>::into_response(http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
     5 [ label = "&app::Singleton"]
-    6 [ label = "<http::Response::<http_body::combinators::BoxBody::<bytes::Bytes, pavex_runtime::Error>> as pavex_runtime::response::IntoResponse>::into_response(http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
-    1 -> 0 [ ]
-    2 -> 1 [ ]
-    3 -> 0 [ ]
-    4 -> 3 [ ]
+    1 -> 0 [ label = "&"]
+    2 -> 0 [ label = "&"]
+    0 -> 4 [ ]
     5 -> 0 [ ]
-    0 -> 6 [ ]
 }
 digraph app_state {
     0 [ label = "crate::ApplicationState(app::Singleton) -> crate::ApplicationState"]
