+++ libs/pavex_cli/tests/ui_tests/route_parameters_happy_path/expectations/diagnostics.dot
@@ -5,19 +5,17 @@ digraph "GET /home/:home_id" {
     3 [ label = "pavex_runtime::extract::route::RawRouteParams<'_, '_>"]
     4 [ label = "<alloc::string::String as pavex_runtime::response::IntoResponse>::into_response(alloc::string::String) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
     5 [ label = "core::prelude::rust_2015::v1::Result<pavex_runtime::extract::route::RouteParams<app::HomeRouteParams>, pavex_runtime::extract::route::errors::ExtractRouteParamsError> -> pavex_runtime::extract::route::errors::ExtractRouteParamsError"]
-    6 [ label = "pavex_runtime::extract::route::errors::ExtractRouteParamsError -> &pavex_runtime::extract::route::errors::ExtractRouteParamsError"]
-    7 [ label = "pavex_runtime::extract::route::errors::ExtractRouteParamsError::into_response(&pavex_runtime::extract::route::errors::ExtractRouteParamsError) -> http::Response<alloc::string::String>"]
-    8 [ label = "<http::Response::<alloc::string::String> as pavex_runtime::response::IntoResponse>::into_response(http::Response<alloc::string::String>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
-    9 [ label = "`match`"]
+    6 [ label = "pavex_runtime::extract::route::errors::ExtractRouteParamsError::into_response(&pavex_runtime::extract::route::errors::ExtractRouteParamsError) -> http::Response<alloc::string::String>"]
+    7 [ label = "<http::Response::<alloc::string::String> as pavex_runtime::response::IntoResponse>::into_response(http::Response<alloc::string::String>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
+    8 [ label = "`match`"]
     1 -> 0 [ ]
-    9 -> 5 [ ]
+    8 -> 5 [ ]
     3 -> 2 [ ]
     0 -> 4 [ ]
-    9 -> 1 [ ]
-    5 -> 6 [ ]
+    8 -> 1 [ ]
+    5 -> 6 [ label = "&"]
     6 -> 7 [ ]
-    7 -> 8 [ ]
-    2 -> 9 [ ]
+    2 -> 8 [ ]
 }
 digraph "GET /home/:home_id/room/:room_id" {
     0 [ label = "app::get_room(pavex_runtime::extract::route::RouteParams<app::RoomRouteParams>) -> alloc::string::String"]
@@ -26,19 +24,17 @@ digraph "GET /home/:home_id/room/:room_id" {
     3 [ label = "pavex_runtime::extract::route::RawRouteParams<'_, '_>"]
     4 [ label = "<alloc::string::String as pavex_runtime::response::IntoResponse>::into_response(alloc::string::String) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
     5 [ label = "core::prelude::rust_2015::v1::Result<pavex_runtime::extract::route::RouteParams<app::RoomRouteParams>, pavex_runtime::extract::route::errors::ExtractRouteParamsError> -> pavex_runtime::extract::route::errors::ExtractRouteParamsError"]
-    6 [ label = "pavex_runtime::extract::route::errors::ExtractRouteParamsError -> &pavex_runtime::extract::route::errors::ExtractRouteParamsError"]
-    7 [ label = "pavex_runtime::extract::route::errors::ExtractRouteParamsError::into_response(&pavex_runtime::extract::route::errors::ExtractRouteParamsError) -> http::Response<alloc::string::String>"]
-    8 [ label = "<http::Response::<alloc::string::String> as pavex_runtime::response::IntoResponse>::into_response(http::Response<alloc::string::String>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
-    9 [ label = "`match`"]
+    6 [ label = "pavex_runtime::extract::route::errors::ExtractRouteParamsError::into_response(&pavex_runtime::extract::route::errors::ExtractRouteParamsError) -> http::Response<alloc::string::String>"]
+    7 [ label = "<http::Response::<alloc::string::String> as pavex_runtime::response::IntoResponse>::into_response(http::Response<alloc::string::String>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
+    8 [ label = "`match`"]
     1 -> 0 [ ]
-    9 -> 5 [ ]
+    8 -> 5 [ ]
     3 -> 2 [ ]
     0 -> 4 [ ]
-    9 -> 1 [ ]
-    5 -> 6 [ ]
+    8 -> 1 [ ]
+    5 -> 6 [ label = "&"]
     6 -> 7 [ ]
-    7 -> 8 [ ]
-    2 -> 9 [ ]
+    2 -> 8 [ ]
 }
 digraph "GET /town/*town" {
     0 [ label = "app::get_town(pavex_runtime::extract::route::RouteParams<app::TownRouteParams<'_>>) -> alloc::string::String"]
@@ -47,19 +43,17 @@ digraph "GET /town/*town" {
     3 [ label = "pavex_runtime::extract::route::RawRouteParams<'_, '_>"]
     4 [ label = "<alloc::string::String as pavex_runtime::response::IntoResponse>::into_response(alloc::string::String) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
     5 [ label = "core::prelude::rust_2015::v1::Result<pavex_runtime::extract::route::RouteParams<app::TownRouteParams<'_>>, pavex_runtime::extract::route::errors::ExtractRouteParamsError> -> pavex_runtime::extract::route::errors::ExtractRouteParamsError"]
-    6 [ label = "pavex_runtime::extract::route::errors::ExtractRouteParamsError -> &pavex_runtime::extract::route::errors::ExtractRouteParamsError"]
-    7 [ label = "pavex_runtime::extract::route::errors::ExtractRouteParamsError::into_response(&pavex_runtime::extract::route::errors::ExtractRouteParamsError) -> http::Response<alloc::string::String>"]
-    8 [ label = "<http::Response::<alloc::string::String> as pavex_runtime::response::IntoResponse>::into_response(http::Response<alloc::string::String>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
-    9 [ label = "`match`"]
+    6 [ label = "pavex_runtime::extract::route::errors::ExtractRouteParamsError::into_response(&pavex_runtime::extract::route::errors::ExtractRouteParamsError) -> http::Response<alloc::string::String>"]
+    7 [ label = "<http::Response::<alloc::string::String> as pavex_runtime::response::IntoResponse>::into_response(http::Response<alloc::string::String>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
+    8 [ label = "`match`"]
     1 -> 0 [ ]
-    9 -> 5 [ ]
+    8 -> 5 [ ]
     3 -> 2 [ ]
     0 -> 4 [ ]
-    9 -> 1 [ ]
-    5 -> 6 [ ]
+    8 -> 1 [ ]
+    5 -> 6 [ label = "&"]
     6 -> 7 [ ]
-    7 -> 8 [ ]
-    2 -> 9 [ ]
+    2 -> 8 [ ]
 }
 digraph app_state {
     0 [ label = "crate::ApplicationState() -> crate::ApplicationState"]
