+++ libs/pavex_cli/tests/ui_tests/trait_methods_are_supported/expectations/diagnostics.dot
@@ -1,24 +1,20 @@
 digraph "GET /home" {
     0 [ label = "app::handler(app::A, app::C, app::D, app::E) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
     1 [ label = "<app::C as app::GenericTrait::<std::string::String>>::a_method(&app::C) -> app::E"]
-    2 [ label = "app::C -> &app::C"]
-    3 [ label = "<app::B as app::AnotherTrait>::a_method_that_consumes_self(app::B) -> app::C"]
-    4 [ label = "<app::A as app::MyTrait>::a_method_that_borrows_self(&app::A) -> app::B"]
-    5 [ label = "app::A -> &app::A"]
-    6 [ label = "<app::A as app::MyTrait>::a_method_that_returns_self() -> app::A"]
-    7 [ label = "<app::A as app::MyTrait>::a_method_with_a_generic::<std::string::String>(&app::A) -> app::D"]
-    8 [ label = "<http::Response::<http_body::combinators::BoxBody::<bytes::Bytes, pavex_runtime::Error>> as pavex_runtime::response::IntoResponse>::into_response(http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
+    2 [ label = "<app::B as app::AnotherTrait>::a_method_that_consumes_self(app::B) -> app::C"]
+    3 [ label = "<app::A as app::MyTrait>::a_method_that_borrows_self(&app::A) -> app::B"]
+    4 [ label = "<app::A as app::MyTrait>::a_method_that_returns_self() -> app::A"]
+    5 [ label = "<app::A as app::MyTrait>::a_method_with_a_generic::<std::string::String>(&app::A) -> app::D"]
+    6 [ label = "<http::Response::<http_body::combinators::BoxBody::<bytes::Bytes, pavex_runtime::Error>> as pavex_runtime::response::IntoResponse>::into_response(http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>) -> http::Response<http_body::combinators::BoxBody<bytes::Bytes, pavex_runtime::Error>>"]
     1 -> 0 [ ]
-    2 -> 1 [ ]
+    2 -> 1 [ label = "&"]
     3 -> 2 [ ]
-    4 -> 3 [ ]
-    5 -> 4 [ ]
-    6 -> 5 [ ]
-    7 -> 0 [ ]
-    5 -> 7 [ ]
-    3 -> 0 [ ]
-    6 -> 0 [ ]
-    0 -> 8 [ ]
+    4 -> 3 [ label = "&"]
+    5 -> 0 [ ]
+    4 -> 5 [ label = "&"]
+    2 -> 0 [ ]
+    4 -> 0 [ ]
+    0 -> 6 [ ]
 }
 digraph app_state {
     0 [ label = "crate::ApplicationState() -> crate::ApplicationState"]
